Abstract order button pair as separate component
