+++ portal/src/DetailsListWithOrdering.tsx
@@ -1,13 +1,8 @@
 import React from "react";
-import {
-  DetailsList,
-  IDetailsListProps,
-  IColumn,
-  IconButton,
-} from "@fluentui/react";
+import { DetailsList, IDetailsListProps, IColumn } from "@fluentui/react";
 import { Context } from "@oursky/react-messageformat";
 
-import styles from "./DetailsListWithOrdering.module.scss";
+import OrderButtons from "./OrderButtons";
 
 interface DetailsListWithOrderingProps extends IDetailsListProps {
   onSwapClicked: (index1: number, index2: number) => void;
@@ -22,79 +17,6 @@ interface DetailsListWithOrderingProps extends IDetailsListProps {
   orderColumnMaxWidth?: number;
 }
 
-interface OrderColumnButtonsProps {
-  index?: number;
-  itemCount: number;
-  onSwapClicked: (index1: number, index2: number) => void;
-  renderAriaLabel: (index?: number) => string;
-}
-
-export const OrderColumnButtons: React.FC<OrderColumnButtonsProps> = function OrderColumnButtons(
-  props: OrderColumnButtonsProps
-) {
-  const { index, itemCount, onSwapClicked, renderAriaLabel } = props;
-  const { renderToString } = React.useContext(Context);
-  const onUpClicked = React.useCallback(() => {
-    if (index == null) {
-      return;
-    }
-    onSwapClicked(index, index - 1);
-  }, [index, onSwapClicked]);
-  const onDownClicked = React.useCallback(() => {
-    if (index == null) {
-      return;
-    }
-    onSwapClicked(index, index + 1);
-  }, [index, onSwapClicked]);
-
-  const ariaLabelUp = React.useMemo(() => {
-    return renderToString("DetailsListWithOrdering.move-up", {
-      key: renderAriaLabel(index),
-    });
-  }, [renderAriaLabel, index, renderToString]);
-  const ariaLabelDown = React.useMemo(() => {
-    return renderToString("DetailsListWithOrdering.move-down", {
-      key: renderAriaLabel(index),
-    });
-  }, [renderAriaLabel, index, renderToString]);
-
-  return (
-    <div>
-      <IconButton
-        className={styles.orderColumnButton}
-        disabled={index === itemCount - 1}
-        onClick={onDownClicked}
-        iconProps={{ iconName: "ChevronDown" }}
-        ariaLabel={ariaLabelDown}
-      />
-      <IconButton
-        className={styles.orderColumnButton}
-        disabled={index === 0}
-        onClick={onUpClicked}
-        iconProps={{ iconName: "ChevronUp" }}
-        ariaLabel={ariaLabelUp}
-      />
-    </div>
-  );
-};
-
-export function swap<T>(items: T[], index1: number, index2: number): T[] {
-  const newItems = [...items];
-  const thisItem = newItems[index1];
-  const thatItem = newItems[index2];
-  if (
-    index1 < 0 ||
-    index2 < 0 ||
-    index1 >= items.length ||
-    index2 >= items.length
-  ) {
-    return items;
-  }
-  newItems[index1] = thatItem;
-  newItems[index2] = thisItem;
-  return newItems;
-}
-
 const DetailsListWithOrdering: React.FC<DetailsListWithOrderingProps> = function DetailsListWithOrdering(
   props: DetailsListWithOrderingProps
 ) {
@@ -103,7 +25,7 @@ const DetailsListWithOrdering: React.FC<DetailsListWithOrderingProps> = function
     (item?: any, index?: number, column?: IColumn) => {
       if (column?.key === "order") {
         return (
-          <OrderColumnButtons
+          <OrderButtons
             index={index}
             itemCount={props.items.length}
             onSwapClicked={props.onSwapClicked}
