+++ portal/src/OrderButtons.module.scss
@@ -1,8 +1,4 @@
-.flip {
-  transform: rotate(180deg);
-}
-
-.orderColumnButton {
+.orderButton {
   padding: 5px;
   margin: 2px 8px 2px 0;
   background: transparent;
